CurriculumVitaeBuilder/CurriculumVitaeBuilder.Domain/Command/CvSections/Contact/Create/CreateContactSectionValidator.cs
// <copyright file="CreateContactSectionValidator.cs" company="BJSS">
// Copyright (c) BJSS. All rights reserved.
// </copyright>

namespace CurriculumVitaeBuilder.Domain.Command.CvSections.Contact.Create
{
    using Chest.Core.Command;

    using FluentValidation;

    /// <summary>
    /// Create Section Validator.
    /// </summary>
    public class CreateContactSectionValidator : CommandValidator<CreateContactSection>
    {
        /// <inheritdoc/>
        public override void ConfigureRules()
        {
            this.RuleFor(x => x.UserId)
                .NotEmpty()
                .WithMessage("A valid user identifier must be provided");

            this.RuleFor(x => x.CvId)
                .NotEmpty()
                .WithMessage("A valid cv identifier must be provided");

            this.RuleFor(x => x.ContactDetails)
                .NotEmpty()
                .NotNull()
                .WithMessage("A contact detail must be provided");
        }
    }
}
